Router, componente dropdown, portfolio401

+++ nextjs-dashboard/src/data/portfolios.ts
@@ -2,28 +2,61 @@ import { Portfolio } from "../types/portfolio";
 import { Investment } from "../types/investment";
 import { Asset } from "../types/asset";
 
-const API_BASE_URL = "http://localhost:8888"; // Change this to your API base URL
+const API_BASE_URL = "http://localhost:8080"; // Change this to your API base URL
+const getTokenFromLocalStorage = () => localStorage.getItem("jwtToken");
+const token1 = getTokenFromLocalStorage();
+const trimmedTokenOut = token1.trim().replace("", );
 
+console.log(token1)
+console.log(trimmedTokenOut)
 // Function to fetch all portfolios
 export const getAllPortfolios = async (): Promise<Portfolio[]> => {
-  const response = await fetch(`${API_BASE_URL}/portfolios/getAll`);
+  const token = getTokenFromLocalStorage();
+  if (!token) {
+    throw new Error("No token found in local storage");
+  }
+  const trimmedToken = token1.trim().replace("", );
+  
+  const response = await fetch(`${API_BASE_URL}/portfolios/getAll`, {
+    method: "GET",
+    headers: {
+      Authorization: `Bearer ${trimmedToken}`,
+      "Content-Type": "application/json",
+    },
+    mode: "no-cors",
+  });
+
   if (!response.ok) {
     throw new Error("Failed to fetch portfolios");
   }
+
   return response.json();
 };
 
 // Function to fetch a portfolio by its ID
 export const getPortfolioById = async (
-  id: number,
+  id: number
 ): Promise<Portfolio | null> => {
-  const response = await fetch(`${API_BASE_URL}/portfolios/${id}`);
+  const token = getTokenFromLocalStorage();
+  if (!token) {
+    throw new Error("No token found in local storage");
+  }
+
+  const response = await fetch(`${API_BASE_URL}/portfolios/${id}`, {
+    headers: {
+      Authorization: `Bearer ${token}`,
+      "Content-Type": "application/json",
+    },
+    mode: "no-cors",
+  });
+
   if (!response.ok) {
     if (response.status === 404) {
       return null; // Portfolio not found
     }
     throw new Error("Failed to fetch portfolio");
   }
+
   return response.json();
 };
 
@@ -32,16 +65,25 @@ export const updatePortfolio = async (
   id: number,
   portfolio: Portfolio,
 ): Promise<Portfolio | null> => {
+  const token = getTokenFromLocalStorage();
+  if (!token) {
+    throw new Error("No token found in local storage");
+  }
+
   const response = await fetch(`${API_BASE_URL}/portfolios/update/${id}`, {
     method: "PUT",
     headers: {
+      Authorization: `Bearer ${token}`,
       "Content-Type": "application/json",
     },
+    mode: "no-cors",
     body: JSON.stringify(portfolio),
   });
+
   if (!response.ok) {
     throw new Error("Failed to update portfolio");
   }
+
   return response.json();
 };
 
@@ -49,24 +91,44 @@ export const updatePortfolio = async (
 export const createPortfolio = async (
   portfolio: Portfolio,
 ): Promise<Portfolio | null> => {
+  const token = getTokenFromLocalStorage();
+  if (!token) {
+    throw new Error("No token found in local storage");
+  }
+
   const response = await fetch(`${API_BASE_URL}/portfolios/create`, {
     method: "POST",
     headers: {
+      Authorization: `Bearer ${token}`,
       "Content-Type": "application/json",
     },
+    mode: "no-cors",
     body: JSON.stringify(portfolio),
   });
+
   if (!response.ok) {
     throw new Error("Failed to create portfolio");
   }
+
   return response.json();
 };
 
 // Function to delete a portfolio
 export const deletePortfolio = async (id: number): Promise<void> => {
+  const token = getTokenFromLocalStorage();
+  if (!token) {
+    throw new Error("No token found in local storage");
+  }
+
   const response = await fetch(`${API_BASE_URL}/portfolios/delete/${id}`, {
     method: "DELETE",
+    headers: {
+      Authorization: `Bearer ${token}`,
+      "Content-Type": "application/json",
+    },
+    mode: "no-cors",
   });
+
   if (!response.ok) {
     throw new Error("Failed to delete portfolio");
   }
@@ -74,10 +136,23 @@ export const deletePortfolio = async (id: number): Promise<void> => {
 
 // Function to fetch all investments
 export const getAllInvestments = async (): Promise<Investment[]> => {
-  const response = await fetch(`${API_BASE_URL}/investments/getAll`);
+  const token = getTokenFromLocalStorage();
+  if (!token) {
+    throw new Error("No token found in local storage");
+  }
+
+  const response = await fetch(`${API_BASE_URL}/investments/getAll`, {
+    headers: {
+      Authorization: `Bearer ${token}`,
+      "Content-Type": "application/json",
+    },
+    mode: "no-cors",
+  });
+
   if (!response.ok) {
     throw new Error("Failed to fetch investments");
   }
+
   return response.json();
 };
 
@@ -85,13 +160,26 @@ export const getAllInvestments = async (): Promise<Investment[]> => {
 export const getInvestmentById = async (
   id: number,
 ): Promise<Investment | null> => {
-  const response = await fetch(`${API_BASE_URL}/investments/${id}`);
+  const token = getTokenFromLocalStorage();
+  if (!token) {
+    throw new Error("No token found in local storage");
+  }
+
+  const response = await fetch(`${API_BASE_URL}/investments/${id}`, {
+    headers: {
+      Authorization: `Bearer ${token}`,
+      "Content-Type": "application/json",
+    },
+    mode: "no-cors",
+  });
+
   if (!response.ok) {
     if (response.status === 404) {
       return null; // Investment not found
     }
     throw new Error("Failed to fetch investment");
   }
+
   return response.json();
 };
 
@@ -100,16 +188,25 @@ export const updateInvestment = async (
   id: number,
   investment: Investment,
 ): Promise<Investment | null> => {
+  const token = getTokenFromLocalStorage();
+  if (!token) {
+    throw new Error("No token found in local storage");
+  }
+
   const response = await fetch(`${API_BASE_URL}/investments/update/${id}`, {
     method: "PUT",
     headers: {
+      Authorization: `Bearer ${token}`,
       "Content-Type": "application/json",
     },
+    mode: "no-cors",
     body: JSON.stringify(investment),
   });
+
   if (!response.ok) {
     throw new Error("Failed to update investment");
   }
+
   return response.json();
 };
 
@@ -117,24 +214,43 @@ export const updateInvestment = async (
 export const createInvestment = async (
   investment: Investment,
 ): Promise<Investment | null> => {
+  const token = getTokenFromLocalStorage();
+  if (!token) {
+    throw new Error("No token found in local storage");
+  }
+
   const response = await fetch(`${API_BASE_URL}/portfolios/addInvestment`, {
     method: "PUT",
     headers: {
+      Authorization: `Bearer ${token}`,
       "Content-Type": "application/json",
     },
+    mode: "no-cors",
     body: JSON.stringify(investment),
   });
+
   if (!response.ok) {
     throw new Error("Failed to create investment");
   }
+
   return response.json();
 };
 
 // Function to delete an investment
 export const deleteInvestment = async (id: number): Promise<void> => {
+  const token = getTokenFromLocalStorage();
+  if (!token) {
+    throw new Error("No token found in local storage");
+  }
+
   const response = await fetch(`${API_BASE_URL}/investments/delete/${id}`, {
     method: "DELETE",
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
+    mode: "no-cors",
   });
+
   if (!response.ok) {
     throw new Error("Failed to delete investment");
   }
@@ -142,9 +258,21 @@ export const deleteInvestment = async (id: number): Promise<void> => {
 
 // Function to fetch all assets
 export const getAllAssets = async (): Promise<Asset[]> => {
-  const response = await fetch(`${API_BASE_URL}/assets/getAll`);
-  if (!response.ok) {
-    throw new Error("Failed to fetch investments");
+  const token = getTokenFromLocalStorage();
+  if (!token) {
+    throw new Error("No token found in local storage");
   }
+
+  const response = await fetch(`${API_BASE_URL}/assets/getAll`, {
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
+    mode: "no-cors",
+  });
+
+  if (!response.ok) {
+    throw new Error("Failed to fetch assets");
+  }
+
   return response.json();
 };
